requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
Markdown==3.4.4
PyJWT==1.7.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==6.6.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
